keydown no longer keeps triggering

+++ src/keyEvents.ts
@@ -3,6 +3,8 @@ import iohook from "iohook";
 
 type Callback = ()=>any;
 
+const keyStates = new Set<number>();
+
 const keyDownListenersStorage = new Map<number, Callback[]>();
 const keyUpListenersStorage = new Map<number, Callback[]>();
 
@@ -15,18 +17,24 @@ function triggerListeners(key:number, storage:Map<number, Callback[]>) {
 }
 
 iohook.addListener("keydown", e=>{
+    if(keyStates.has(e.rawcode)) return;
+    keyStates.add(e.rawcode);
     triggerListeners(e.rawcode, keyDownListenersStorage);
 });
 
 iohook.addListener("keyup", e=>{
+    keyStates.delete(e.rawcode);
     triggerListeners(e.rawcode, keyUpListenersStorage);
 });
 
 iohook.addListener("mousedown", e=>{
+    if(keyStates.has(-e.button)) return;
+    keyStates.add(-e.button);
     triggerListeners(-e.button, keyDownListenersStorage);
 });
 
 iohook.addListener("mouseup", e=>{
+    keyStates.delete(-e.button);
     triggerListeners(-e.button, keyUpListenersStorage);
 });
 
@@ -41,6 +49,10 @@ export function addKeyDownListener(key:number, callback:Callback) {
 }
 export function addKeyUpListener(key:number, callback:Callback) {
     addKeyListener(key, callback, keyUpListenersStorage);
+}
+
+export function isKeyDown(key:number):boolean {
+    return keyStates.has(key);
 }
 
 function removeKeyListener(key:number, callback:Callback, storage:Map<number, Callback[]>) {
@@ -58,7 +70,14 @@ export function removeKeyUpListener(key:number, callback:Callback) {
     removeKeyListener(key, callback, keyUpListenersStorage);
 }
 
-export function removeKeyListeners(key:number) {
-    keyUpListenersStorage.delete(key);
-    keyDownListenersStorage.delete(key);
+export function removeKeyListeners(key?:number) {
+    if(key == null) {
+        keyUpListenersStorage.clear();
+        keyDownListenersStorage.clear();
+    }
+    else {
+        keyUpListenersStorage.delete(key);
+        keyDownListenersStorage.delete(key);
+    }
+
 }
